Renamed purelypeer app to vouchers
