--- purelypeer/consumers/__init__.py
@@ -1,1 +1,0 @@
-from purelypeer.consumers.voucher_consumer import *
