+++ requirements.txt
@@ -24,6 +24,7 @@ django-jinja==2.6.0
 django-picklefield==3.0.1
 django-postgres-extra==2.0.2
 django-push-notifications==3.0.0
+django-rename-app==0.1.6
 django-rest-hooks==1.5.0
 djangorestframework==3.11.0
 djangorestframework-simplejwt==4.4.0
