+++ vouchers/consumers/__init__.py
@@ -1,0 +1,1 @@
+from vouchers.consumers.voucher_consumer import *
