+++ vouchers/js/src/funcs/utils.js
