+++ vouchers/vault.py
@@ -2,8 +2,8 @@ from django.conf import settings
 
 from paytacapos.models import Merchant
 
-from purelypeer.js.runner import ScriptFunctions
-from purelypeer.models import Vault, Voucher
+from vouchers.js.runner import ScriptFunctions
+from vouchers.models import Vault, Voucher
 
 from main.utils.subscription import new_subscription
 from main.models import Address
